# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get home feed
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /posts/explore:
    get:
      operationId: getExplorePosts
      tags: [posts]
      summary: Get explore feed
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Explore posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePost
      tags: [posts]
      summary: Edit post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /posts/{postId}/like:
    post:
      operationId: togglePostLike
      tags: [posts]
      summary: Like or unlike a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/user/{userId}:
    get:
      operationId: getUserPosts
      tags: [posts]
      summary: Get all posts by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"

  # COMMENTS
...
  /messages/conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get all conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /messages/conversations/{conversationId}:
    get:
      operationId: getConversationMessages
      tags: [messages]
      summary: Get messages in conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        profilePicture:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isFollowing:
          type: boolean
        followersCount:
          type: integer
      required: [id, username, displayName, isFollowing, followersCount]
...
    PostMedia:
      type: object
      properties:
        url:
          type: string
        type:
          type: string
          enum: [image, video]
      required: [url, type]
    Post:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
        author:
          $ref: "#/components/schemas/UserSummary"
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        visibility:
          type: string
          enum: [public, followers, private]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, content, media, author, likesCount, commentsCount, isLiked, visibility, createdAt, updatedAt]
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        totalCount:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
      required: [posts, totalCount, page, hasMore]
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
        visibility:
          type: string
          enum: [public, followers, private]
          default: public
      required: [content]
    UpdatePostBody:
      type: object
      properties:
        content:
          type: string
        visibility:
          type: string
          enum: [public, followers, private]
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required: [liked, likesCount]

    # COMMENTS
...
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        sender:
          $ref: "#/components/schemas/UserSummary"
        text:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        readBy:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, sender, readBy, createdAt]
    Conversation:
      type: object
      properties:
        id:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        unreadCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required: [id, participants, unreadCount, updatedAt]
    SendMessageBody:
      type: object
      properties:
        recipientId:
          type: string
        text:
          type: string
        mediaUrl:
          type: string
      required: [recipientId]